fix: Resolve SwiftUI namespace conflicts and refinements
- Fix Circle() → SwiftUI.Circle() to avoid conflicts with custom Circle types
- Apply consistent namespace disambiguation across sync, celebration, and calendar components
- Refinements to splash screen, task detail, and momentum views

Files updated:
- Sync indicators (OfflineBanner, SyncStatusIndicator)
- Calendar components (CosmosTimeBlock, NowHorizonIndicator, CosmosWeekView)
- Celebration views (DailyWins, ShareableWinCard, StreakMomentum)
- Momentum dashboard and stats components
- Cosmic splash screen
- Oracle task detail sheet

+++ Veloce/Services/CelebrationSounds.swift
@@ -301,10 +301,10 @@ final class ToneSynthesizer {
             let decayStart = duration * 0.7
 
             if time < attackTime {
-                sample *= Float(time / attackTime)
+                sample *= time / attackTime
             } else if time > decayStart {
                 let decayProgress = (time - decayStart) / (duration - decayStart)
-                sample *= Float(1.0 - decayProgress)
+                sample *= 1.0 - decayProgress
             }
 
             data[i] = Float(sample) * volume
